+++ R/print.netmeta.R
@@ -441,6 +441,7 @@ print.netmeta <- function(x,
       ##
       tmat <- data.frame(abbr, full)
       names(tmat) <- c("Abbreviation", "Treatment name")
+      tmat <- tmat[abbr != full, ]
       tmat <- tmat[order(tmat$Abbreviation), ]
       ##
       if (legend) {