+++ R/rankogram.R
@@ -362,10 +362,12 @@ print.rankogram <- function(x,
       trts <- rownames(x$ranking.matrix.random)
     ##
     trts.abbr <- treats(trts, nchar.trts)
-    if (any(trts != trts.abbr)) {
+    diff.trts <- trts != trts.abbr
+    if (any(diff.trts)) {
       cat("\nLegend:\n")
       tmat <- data.frame(trts.abbr, trts)
       names(tmat) <- c("Abbreviation", "Treatment name")
+      tmat <- tmat[diff.trts, ]
       tmat <- tmat[order(tmat$Abbreviation), ]
       ##
       prmatrix(tmat, quote = FALSE, right = TRUE,