+++ R/summary.netmeta.R
@@ -1017,6 +1017,7 @@ print.summary.netmeta <- function(x,
     ##
     tmat <- data.frame(abbr, full)
     names(tmat) <- c("Abbreviation", "Treatment name")
+    tmat <- tmat[abbr != full, ]
     tmat <- tmat[order(tmat$Abbreviation), ]
     ##
     if (legend) {